wip clean up and restructure
